Cast to size_t to remove warning. Cvs:
	* gnuserv.c: (permitted): Cast to size_t to remove warning.
Cvs: ----------------------------------------------------------------------

+++ src/daemon/gnuserv.c
@@ -203,7 +203,7 @@ permitted (u_long host_addr, int fd)
 
 	auth_data_len = atoi (buf);
 	
-	if (auth_data_len < 1 || auth_data_len > sizeof(buf)) {
+	if (auth_data_len < 1 || (size_t)auth_data_len > sizeof(buf)) {
 	    syslog_message(LOG_WARNING, "Invalid data length supplied by client");
 	    return FALSE;
 	}
